Able to start and stop process using admin_state toggling

+++ agent/src/github.com/jetblack87/maestro/agent/data/structs.go
@@ -32,5 +32,5 @@ type Process struct {
 	ProcessClass string
 	AdminState string
 	OperState string
-	Pid int64
+	Pid int
 }
